// src/app/models/test.model.ts
export interface Test{
    parteA_arr: Object; 
    parteB_arr: Object;
    parteC_arr:  Object; 
    parteD_arr: Object;
    parteActividades_arr: Object; 
    parteHabilidades_arr: Object; 
    parteOcupaciones_arr: Object; 
    auto1_arr: Object; 
    auto2_arr: Object; 
    total_arr: Object; 
}