// src/app/models/encuesta.model.ts
export interface Encuesta{
    nombreInstitucion: String;
    estratoSocioeconomico: String;
    genero: String;
    rangoEdades: String; 
    preferenciaEstudiosSuperiores: String;
    asignaturaFilosofiaYLiteratura: String; 
    asignaturaMatematicas: String;
    asignaturaSocialesYCiudadanas: String; 
    asignaturaCienciasNaturales: String; 
    asignaturaIdiomas: String; 
    asignaturaDeportes: String; 
    asignaturaArtesYDanza: String;
}